% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_bimodal}
\alias{modis_bimodal}
\title{Check for Bimodality in MODIS Band}
\usage{
modis_bimodal(x = NULL, watermask = NULL, drymask = NULL)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{watermask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the water-polygon that is used to extract reflectances of water
on MODIS band 7.}

\item{drymask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the dryland-polygon that is used to extract reflectances of
water on MODIS band 7.}
}
\value{
logical indicating if the image is bimodal (TRUE) or not bimodal
(FALSE)
}
\description{
Function to check for bimodality in MODIS band 7. For details check Wolksi et
al. 2017.
}
