% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_classify}
\alias{modis_classify}
\title{Classify MODIS Image}
\usage{
modis_classify(
  x = NULL,
  watermask = NULL,
  drymask = NULL,
  ignore.bimodality = F
)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{watermask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the water-polygon that is used to extract reflectances of water
on MODIS band 7.}

\item{drymask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the dryland-polygon that is used to extract reflectances of
water on MODIS band 7.}

\item{ignore.bimodality}{logical. Should the bimodality check be ignored?}
}
\value{
\code{RasterLayer} of classified MODIS image
}
\description{
Function to classify a MODIS image into the binary categories water and
dryland. Water = 1, Dryland = 0.
}
