% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_date}
\alias{modis_date}
\title{Retrieve Date from MODIS Filenames}
\usage{
modis_date(filename = NULL)
}
\arguments{
\item{filename}{character. Filename(s) of the files for which a date should
be extracted}
}
\value{
character. Dates as derived from the MODIS filename
}
\description{
Function to retrieve the date from a MODIS hdf file. Note that this function
only works with the original MODIS hdf files.
}
