% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_download}
\alias{modis_download}
\title{Download MODIS products}
\usage{
modis_download(
  dataframe = NULL,
  path = getwd(),
  username = NULL,
  password = NULL,
  overwrite = F
)
}
\arguments{
\item{dataframe}{\code{data.frame}. Selection of the files that should be
downloaded (see \code{modis_files()})}

\item{path}{character. Path were the downloaded files should be stored to}

\item{username}{character. Username of your EarthData account}

\item{password}{character. Password of your EarthData account}

\item{overwrite}{logical. If such a file already exists in the path, should
it be overwritten?}
}
\value{
Character vector of filename pointing to the downloaded files
}
\description{
This function uses the data from the \code{modis_files()} function to
download the desired MODIS files.
}
