% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_extract}
\alias{modis_extract}
\title{Extract Modis Bands and Convert HDF to Raster}
\usage{
modis_extract(
  filepath = NULL,
  outdir = dirname(filepath),
  removeHDF = F,
  overwrite = F
)
}
\arguments{
\item{filepath}{character. Filepath pointing to the downloaded hdf files.
Note that the filenames need to be unchanged and as downloaded from using
\code{modis_download()}}

\item{outdir}{character. Directory to which the converted file should be
stored. By default it is stored in the same folder as the hdf file.}

\item{removeHDF}{logical. Should the original hdf file be removed after
conversion?}
}
\value{
filepath to the converted file
}
\description{
Function to convert the downloaded hdf files to proper .tif rasters
}
