% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_files}
\alias{modis_files}
\title{Get Available Files of a MODIS Product}
\usage{
modis_files(
  product = "MCD43A4",
  version = "006",
  start_date = NULL,
  end_date = NULL,
  aoi = NULL,
  limit = 100
)
}
\arguments{
\item{product}{character. Available products can be shown using
\code{modis_products}}

\item{version}{character. Indicates which version of the product should be
used. By default this is set to "006"}

\item{start_date}{character. Start date for the data requested formatted
yyyy-mm-dd}

\item{end_date}{character. End date for the data requested formatted
yyyy-mm-dd}

\item{aoi}{Any object from which an extent (using \code{bbox()} can be
derived)}

\item{limit}{positive integer. Number of results collected before the process
is aborted}
}
\value{
\code{data.frame} containing some information about all files that
can be downloaded. The column "Online Access URLs" can be used in the
\code{modis_download()} function
}
\description{
Function to retrieve available files from a modis product. This function
allows you to find and select files that you want to download later using
\code{modis_download}.
}
\examples{
files <- modis_files(product = "MCD43A4", version = "006", start_date =
"2020-02-01", end_date = "2020-02-04", aoi = '21.75,-20.65,24.3,-18.15',
limit = 100)
}
