% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_products}
\alias{modis_products}
\title{Show available MODIS products}
\usage{
modis_products(product)
}
\arguments{
\item{product}{character indicating the desired products}
}
\value{
\code{data.frame} containing all available products that fulfill the
search requirement. The column \code{short_name} can be used to retrieve
available files using the \code{modis_files()} function.
}
\description{
Function to show available products. This function allows you to determine
which MODIS products are available for download.
}
\examples{
modis_products("MCD43A4")
}
