% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_specs}
\alias{modis_specs}
\title{Compare Spectral Signatures of MODIS Maps}
\usage{
modis_specs(x = NULL, water = NULL, dryland = NULL)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{water}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent waters}

\item{dryland}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent dryland}
}
\value{
Plot of spatial reflectance values below the two polygons
}
\description{
Function to compare spectral signatures of MODIS maps below water and dryland
polygons
}
