% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_stitch}
\alias{modis_stitch}
\title{Stitch Modis Tiles}
\usage{
modis_stitch(filepaths, outdir = getwd(), outname = NULL, overwrite = F)
}
\arguments{
\item{filepaths}{character. Filepaths pointing to the tiff files as prepared
using the \code{modis_extract} function. The selected files will be stitched}

\item{outname}{character. Filepath and filename to which the stitched tiff
should be stored}
}
\value{
character pointing to the stitched tiff file
}
\description{
Function that allows you to stitch modis tiles together
}
