% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_watermask}
\alias{modis_watermask}
\title{Create Dynamic Watermask}
\usage{
modis_watermask(
  date = NULL,
  filenames = NULL,
  filedates = NULL,
  years = 5,
  threshold = 0.99
)
}
\arguments{
\item{date}{date. Date for which a watermask should be calculated}

\item{filedates}{dates vector. Dates which the above floodmaps represent}

\item{years}{numeric. Number of years that should be considered to create a
watermask}

\item{threshold}{numeric. Needs to be between 0 and 1. How often (relative
frequency) a pixel needs to be inundated in the past x years to be considered
in the watermask}

\item{floodmaps}{character vector. Filenames of all floodmaps that already
exist}
}
\value{
\code{SpatialPolygons} of the waterask
}
\description{
Function to create a dynamic watermask for a point in time based on previous
watermaps. The function uses floodmaps from previous dates to determine areas
that were covered by water in at least x\% of the time. This can help to deal
with issues of non-bimodality.
}
