% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modisDate}
\alias{modisDate}
\title{Retrieve Date from MODIS Filenames}
\usage{
modisDate(filename)
}
\arguments{
\item{filename}{character. Filename(s) of the files for which a date should
be extracted}
}
\value{
character. Dates as derived from the MODIS filename
}
\description{
Function that allows you to retrieve the date from a MODIS file
}
