% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_classify}
\alias{modis_classify}
\title{Classify MODIS Image}
\usage{
modis_classify(x, water, dryland)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{water}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent waters}

\item{dryland}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent dryland}
}
\value{
\code{RasterLayer} of classified MODIS image
}
\description{
Function to classify a MODIS image into the binary categories water (value =
\enumerate{
\item and dryland (value = 0)
}
}
