% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_classify}
\alias{modis_classify}
\title{Classify MODIS Image}
\usage{
modis_classify(
  x = NULL,
  watermask = NULL,
  drymask = NULL,
  ignore.bimodality = F
)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{watermask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the water-polygon that is used to extract reflectances of water
on MODIS band 7. By default the masks from Wolski et al., 2017 are used. See
also \code{modis_watermask()}.}

\item{drymask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the dryland-polygon that is used to extract reflectances of
water on MODIS band 7. By default the masks from Wolski et al., 2017 are
used.}

\item{ignore.bimodality}{logical. Should issues with bimodality be ignored,
i.e. the bimodality check be skipped? This can lead to biased
classifications but may help in detecting issues.}
}
\value{
\code{RasterLayer} of classified MODIS image. water is valued 1,
dryland valued 0. If there are clouds, they are masked as NA.
}
\description{
Function to classify MODIS MCD43A4 band 7 into the binary categories water
and dryland. Classification is based on the algorithm described in Wolski et
al., 2017 and requires that reflectance values of water- and dryland are
sufficiently distinct. The final map binarily depicts water = 1 and dryland =
0.
}
\examples{
\dontrun{
# Download files for two dates
files <- modis_download(
    dates     = c("2020-01-01", "2020-01-01")
  , outdir    = getwd()
  , tmpdir    = tmpdir()
  , username  = "username"
  , password  = "password"
  , overwrite = F
)

# Load one of them
modis <- modis_load(files[1])

# Classify it
classified <- modis_classify(modis)

# Visualize
plot(classified)
}
}
