% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_download}
\alias{modis_download}
\title{Download MODIS products}
\usage{
modis_download(
  dates = NULL,
  outdir = getwd(),
  tmpdir = tempdir(),
  username = NULL,
  password = NULL,
  overwrite = F
)
}
\arguments{
\item{outdir}{character. Directory were the processed files should be stored
to. This is the raw data that will be used for floodmapping}

\item{username}{character. Username of your EarthData account}

\item{password}{character. Password of your EarthData account}

\item{overwrite}{logical. If such a file already exists in the path, should
it be overwritten?}

\item{date}{character vector of dates. Dates for which a modis image should be
downloaded. Needs to be of format "YYYY-mm-dd".}

\item{tempdir}{character. Directory were intermediary files should be stored.
By default this is determined by \code{tempdir()}. However, you can change
this to another folder.}
}
\value{
Character vector of filename pointing to the downloaded files
}
\description{
This function is used to download and process modis files that can then be
classified into floodmaps using \code{modis_classify()}. Important: MODIS
bands 1-7 will be downloaded, yet only band 7 should be used in
\code{modis_classify()}. Note that the filenames simply correspond to the
aqcuisition date of the MODIS satellite imagery.
}
