% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_load}
\alias{modis_load}
\title{Load Downloaded MODIS Data}
\usage{
modis_load(filepath = NULL)
}
\arguments{
\item{filepath}{character. Filepath pointing to the downloaded modis file}
}
\value{
\code{RasterLayer} containing only band 7 of the downloaded modis
satellite imagery
}
\description{
Function to load a file that was downloaded using \code{modis_download()}.
This function simply wraps the command \code{raster(filepath, band = 7)}.
}
