% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_load}
\alias{modis_load}
\title{Load Downloaded MODIS Data}
\usage{
modis_load(filepath = NULL)
}
\arguments{
\item{filepath}{character. Filepath pointing to the downloaded modis file}
}
\value{
\code{RasterLayer} containing only MODIS MCD43A4 band 7
}
\description{
Function to load downloaded MODIS MCD43A4 band 7 (see
\code{modis_download()}). This function simply wraps the command
\code{raster(filepath, band = 7)}.
}
\examples{
\dontrun{
# Download files for two dates
files <- modis_download(
    dates     = c("2020-01-01", "2020-01-01")
  , outdir    = getwd()
  , tmpdir    = tempdir()
  , username  = "username"
  , password  = "password"
  , overwrite = F
)

# Load one of them and show it
modis <- modis_load(files[1])
show(modis)
}
}
