% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_percentiles}
\alias{modis_percentiles}
\title{Calculate Reflectance Percentiles}
\usage{
modis_percentiles(x = NULL, watermask = NULL, drymask = NULL)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{watermask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the water-polygon that is used to extract reflectances of water
on MODIS band 7. By default the masks from Wolski et al., 2017 are used. See
also \code{modis_watermask()} in case you want to create dynamic watermasks.}

\item{drymask}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing the dryland-polygon that is used to extract reflectances of
water on MODIS band 7. By default the masks from Wolski et al., 2017 are
used.}
}
\value{
dataframe of percentiles
}
\description{
Function to calculate percentiles of water and dryland reflectance values
}
\examples{
\dontrun{
# Download file
file <- modis_download(
    dates     = "2020-01-01"
  , outdir    = getwd()
  , tmpdir    = tempdir()
  , username  = "username"
  , password  = "password"
  , overwrite = F
)

# Load it
modis <- modis_load(file)

# Check percentiles
modis_percentiles(modis)
}
}
