% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_percentiles}
\alias{modis_percentiles}
\title{Calculate Reflectance Percentiles}
\usage{
modis_percentiles(x, water, dryland)
}
\arguments{
\item{x}{\code{RasterLayer} of MODIS band 7}

\item{water}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent waters}

\item{dryland}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
representing permanent dryland}
}
\value{
dataframe of percentiles
\dontrun{
# Download file
file <- modis_download(
    dates     = "2020-01-01"
  , outdir    = getwd()
  , tmpdir    = tmpdir()
  , username  = "username"
  , password  = "password"
  , overwrite = F
)

# Load it
modis <- modis_load(file)

# Check percentiles
modis_percentiles(modis)
}
}
\description{
Function to calculate percentiles of water and dryland reflectance values
}
