% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{modis_watermask}
\alias{modis_watermask}
\title{Create Dynamic Watermasks}
\usage{
modis_watermask(
  date = NULL,
  filenames = NULL,
  filedates = NULL,
  years = 5,
  threshold = 0.99
)
}
\arguments{
\item{date}{date. Date for which a watermask should be calculated}

\item{filedates}{vector of dates. Dates which the above floodmaps represent}

\item{years}{numeric. Number of years that should be considered to create a
watermask}

\item{floodmaps}{character. Filenames of all floodmaps that already exist}

\item{Threshold}{numeric between 0 and 1. How often (relative frequency) a
pixel needs to be inundated in the past x years to be considered in the
watermask}
}
\value{
\code{SpatialPolygons} of the waterask
}
\description{
Function to createa dynamic watermask based on previous floodmaps.
}
